+++ lib/posters/index.js
@@ -1,0 +1,5 @@
+import microblog from './microblog.js'
+
+export default {
+    mb: microblog,
+}
